requirements.txt
smolagents
openai
python-dotenv